+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.5.0-rc.7] - 2023-04-21
+### Added
+- `@rarimo/nft-checkout`: `NFTCheckoutOperation.getDestinationTx` method
+- `@rarimo/nft-checkout`: `@distributedlab/jac` dependency to perform JSON API requests
+
+### Changed
+- `root` - SDK now required Node.js version >= 18
+
 ### Fixed
 - `@rarimo/react-nft-checkout`: Fetching token in the dev mode
 
@@ -154,7 +163,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Under the hood changes
 - Initiated repo with `@rarimo/provider` and `@rarimo/nft-checkout` packages
 
-[Unreleased]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.6...HEAD
+[Unreleased]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.7...HEAD
+[1.5.0-rc.7]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.6...1.5.0-rc.7
 [1.5.0-rc.6]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.5...1.5.0-rc.6
 [1.5.0-rc.5]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.4...1.5.0-rc.5
 [1.5.0-rc.4]: https://github.com/rarimo/js-sdk/compare/1.5.0-rc.3...1.5.0-rc.4